apply lots of updates
