+++ .github/workflows/ci.yml
@@ -1,0 +1,42 @@
+name: CI
+
+on:
+  push:
+    tags: ["*"]
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  test:
+    runs-on: ${{ matrix.platform }}
+    defaults:
+      run:
+        shell: bash -el {0} # needed when using setup-miniconda
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
+        platform: [ubuntu-latest, macos-latest, windows-latest]
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+          mamba-version: "*"
+          channels: funkelab,gurobi,conda-forge
+          channel-priority: true
+
+      - name: install build deps
+        run: mamba install ilpy
+
+      - name: install package
+        run: |
+          python -m pip install -U pip
+          python -m pip install -e .[dev]
+
+      - name: run tests
+        run: pytest --color yes -v
