+++ Makefile
@@ -1,10 +1,8 @@
 default:
-	pip install -r requirements.txt
 	pip install .
 
 install-dev:
-	pip install -r requirements_dev.txt
-	pip install -e .
+	pip install -e .[dev]
 
 .PHONY: tests
 tests:
