+++ README.md
@@ -1,0 +1,8 @@
+# structsvm
+
+[![License](https://img.shields.io/pypi/l/structsvm.svg?color=green)](https://github.com/funkelab/structsvm/raw/main/LICENSE)
+[![PyPI](https://img.shields.io/pypi/v/structsvm.svg?color=green)](https://pypi.org/project/structsvm)
+[![Python Version](https://img.shields.io/pypi/pyversions/structsvm.svg?color=green)](https://python.org)
+[![CI](https://github.com/funkelab/structsvm/actions/workflows/ci.yml/badge.svg)](https://github.com/funkelab/structsvm/actions/workflows/ci.yml)
+
+Structured Bundle Method for Risk Minimization.
