+++ pyproject.toml
@@ -1,0 +1,29 @@
+# https://peps.python.org/pep-0517/
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+# https://peps.python.org/pep-0621/
+[project]
+name = "structsvm"
+description = "Structured Bundle Method for Risk Minimization."
+readme = "README.md"
+requires-python = ">=3.8"
+license = { text = "MIT" }
+authors = [
+    { email = "funkej@janelia.hhmi.org", name = "Jan Funke" },
+]
+dynamic = ["version"]
+dependencies = ['ilpy', 'numpy']
+
+[tool.hatch.version]
+path = "structsvm/__init__.py"
+
+# extras
+# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
+[project.optional-dependencies]
+test = ["pytest>=6.0", "pytest-cov", "flake8"]
+
+[project.urls]
+homepage = "https://github.com/funkelab/structsvm"
+repository = "https://github.com/funkelab/structsvm"
