--- requirements_dev.txt
@@ -1,3 +1,0 @@
-flake8
-pytest
-pytest-cov
