--- setup.py
@@ -1,14 +1,0 @@
-from setuptools import setup
-
-setup(
-        name='structsvm',
-        version='0.1',
-        description='Structured Bundle Method for Risk Minimization.',
-        url='https://github.com/funkelab/structsvm',
-        author='Jan Funke',
-        author_email='funkej@janelia.hhmi.org',
-        license='MIT',
-        packages=[
-            'structsvm'
-        ]
-)
