+++ structsvm/__init__.py
@@ -1,3 +1,6 @@
-from .bundle_method import BundleMethod  # noqa
-from .soft_margin_loss import SoftMarginLoss  # noqa
-from .hamming_costs import HammingCosts  # noqa
+from .bundle_method import BundleMethod
+from .soft_margin_loss import SoftMarginLoss
+from .hamming_costs import HammingCosts
+
+__version__ = '0.1.0'
+__all__ = ['BundleMethod', 'SoftMarginLoss', 'HammingCosts']
